Turn LED on and off with button push
Correctly sets the state so that we toggle the LED on upon button push.
We do not change state again until after the button has been released.
Once button has been released, we can go into a changableState so that
the next time the button is pushed, the LED will change depending upon
the current state of the LED.